add CSVViewTable component and use module.css
