--- src/pages/UploadCSVView/UploadCSVView.css
@@ -1,59 +1,0 @@
-
-.app-container {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 0px 20px;
-  margin: auto 20px;
-}
-
-.csv-table {
-  border-collapse: separate;
-  width: 100%;
-  border-spacing: 0px 20px;
-}
-
-th,
-td {
-  text-align: left;
-  padding: 0 8px;
-  width: 300px;
-  font-size: 32px;
-}
-
-td {
-  padding: 8px;
-  height: 70px;
-}
-
-th:first-child {
-  padding-left: 20px;
-}
-
-td:first-child {
-  padding-left: 20px;
-}
-
-th {
-  background-color: none;
-  font-weight: 500;
-}
-
-td {
-  background-color: #efefef;
-}
-
-.upload-csv-view-form {
-  display: flex;
-  gap: 5px;
-}
-
-.upload-csv-view-form td:last-child {
-  display: flex;
-  /* justify-content: space-evenly; */
-  gap: 20px;
-}
-
-.upload-csv-view-form * {
-  font-size: 1rem;
-}
